fix: nginx cannot dynamic if using hostname
nginx need a resolver to dynamicly resolve hostname. the issue is nerdctl does not provide dns resovler thus when any upstrema down then nginx is down. we do not want that. current available soultion is to use static ip.

the uther solution also is to remove the use of upstream. some cases the upstream is error due to duplicate name even tho on different file.